requirements.txt

requests
python-whois
python-nmap
beautifulsoup4
